requirements.txt
fastapi==0.135.2
uvicorn==0.23.2
sqlalchemy==2.0.48
pydantic==2.10.6
langchain-google-genai==2.0.11
langgraph==0.2.74
python-dotenv==1.0.0
mcp==1.2.1
duckduckgo-search==7.3.2
wikipedia-api==0.13.0
google-adk==1.28.1
google-cloud-aiplatform==1.145.0
google-cloud-logging==3.15.0
google-cloud-storage==3.10.1
google-cloud-secret-manager==2.27.0
psycopg2-binary==2.9.9
